Create Number 1.
This is my solution to assignment 1, it is a function that takes in two parameters that are strings, then uses the Number.isInteger to check if the strings1 and string 2 are divisible by 3 and 5 respectively returning each string if this condition is true, if both conditions are true, it returns a concated string of string1 and string 2. There is a simpler method by using the % that returns the reminder of a division.